CMA-14 cleaning code, adjustments to existing logic, initial implementation of Pit tax deduction system.

+++ accounting-service/src/main/java/org/gminds/accounting_service/service/taxes/TaxCalculator.java
@@ -2,6 +2,8 @@ package org.gminds.accounting_service.service.taxes;
 
 import org.gminds.accounting_service.model.Transaction;
 
+import java.math.BigDecimal;
+
 public interface TaxCalculator<T extends Transaction> {
-    Double calculateTax(T transaction);
+    BigDecimal calculateTax(T transaction);
 }
